requirements.txt
requests==2.28.1
pandas==1.4.3
git+https://github.com/twintproject/twint.git@origin/master#egg=twint
selenium==4.3.0
webdriver_manager==3.8.2
beautifulsoup4==4.11.1
gspread==5.4.0
python-dotenv==0.20.0
gspread==5.4.0
gspread_dataframe==3.3.0
google_api_core==2.8.2
google_api_python_client==2.55.0
google_auth==2.9.1
google_auth_httplib2==0.1.0
google_auth_oauthlib==0.5.2
googleapis_common_protos==1.56.4
